Prototype/Proof of concept Multiplayer Implementation
This is a barebones version that will bring up a widget upon the player pression a button. After which a small widget asking to either create or join a session pops up. Upon choosing create, the widget will disappear and who ever presses join session will join the session after a delay. (As long as they are on the same LAN)