2D map generation
I have done making the 2D map sidescroll version. The rooms are not well made so I did not bother to design, but the map is procedurally generated. If anything that needs to be change of the room design, I can still do so. But right now, since we are firstly focusing on 3D part, I did not bother designing the rooms, and also the rooms are somewhat big. I have also include the dungeon architect plug in in git hub.